@@ -6,7 +6,7 @@ homepage: https://github.com/EdvaldoMartins/flutter_package_pub_dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -16,53 +16,50 @@ dependencies:
 
   #ui
   cupertino_icons: ^0.1.3
-#  flutter_markdown: ^0.3.4
-  flutter_markdown: ^0.5.2
-  flutter_svg: ^0.19.0
+  flutter_markdown: ^0.6.2
+  flutter_svg: ^0.22.0
   html: ^0.14.0+2
-  pull_to_refresh: 1.6.3
+  pull_to_refresh: ^2.0.0
   customized:
     git:
       url: https://github.com/EdvaldoMartins/customized
-      ref: models
 
   #state
-  flutter_mobx: ^1.1.0
-  mobx: ^1.2.1
+  mobx: ^2.0.1
+  flutter_mobx: ^2.0.0
 
   #data class
-  equatable: ^1.0.1
-  freezed_annotation: ^0.12.0
-  json_annotation: ^3.1.1
+  freezed_annotation: ^0.14.1
+  freezed: ^0.14.1+3
+  equatable: ^2.0.2
 
   #network
-  connectivity: ^0.4.8+6
-  dio: ^3.0.10
+  dio: ^4.0.0
+  http: ^0.13.0
+  url_launcher: ^6.0.0-nullsafety
   data_connection_checker: ^0.3.4
-  http: ^0.12.1
 
   # Dependency Injection
-  get_it: ^5.0.3
-  injectable: ^1.0.5
+  get_it: ^7.1.2
 
   #utils
   fluttertoast: ^7.1.5
-  share: ^0.6.3+5
-  url_launcher: ^5.4.2
-  dartz: ^0.9.1
-  intl: ^0.16.0
+  share: ^2.0.4
+  dartz: ^0.10.0-nullsafety.1
+  intl: ^0.17.0
+
+  # Testing
+  mockito: ^5.0.10
 
 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mobx_codegen: ^1.1.2
-  build_runner: ^1.10.9
-  freezed: ^0.12.3
-  injectable_generator: ^1.0.6
-  json_serializable: ^3.5.1
-  build_resolvers: ^1.4.4
+  mobx_codegen: ^2.0.1+3
+  build_runner: ^2.0.1
+  json_serializable: ^4.1.1
+  build_resolvers: ^2.0.1
 
 
 # For information on the generic Dart part of this file, see the
@@ -84,34 +81,6 @@ flutter:
     - assets/icons/
     - assets/images/
 
-  # To add assets to your application, add an assets section, like this:
-  # assets:
-  #  - images/a_dot_burr.jpeg
-  #  - images/a_dot_ham.jpeg
 
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
 #    flutter packages pub run build_runner build --delete-conflicting-outputs
+
