@@ -12,15 +12,14 @@ class DetailPackageViewModel = _DetailPackageViewModel
     with _$DetailPackageViewModel;
 
 abstract class _DetailPackageViewModel extends BaseViewModel with Store {
-  _DetailPackageViewModel(this._repository, this._advancedService)
-      : assert(_repository != null);
+  _DetailPackageViewModel(this._repository, this._advancedService);
 
   final IPackageRepository _repository;
 
   final IAdvancedService _advancedService;
 
   @observable
-  Package _package;
+  Package? _package;
 
   @observable
   Score _score = Score.default_;
@@ -32,13 +31,13 @@ abstract class _DetailPackageViewModel extends BaseViewModel with Store {
   bool _loadingReadme = false;
 
   @observable
-  RequestFailure failure = null;
+  RequestFailure? failure = null;
 
   @computed
   bool get hasError => failure != null;
 
   @computed
-  Package get package => _package;
+  Package get package => _package!;
 
   @computed
   Score get score => _score;
@@ -99,7 +98,7 @@ abstract class _DetailPackageViewModel extends BaseViewModel with Store {
         gitPath: package.latest.pubspec.homepage);
     _loadingReadme = false;
     response.fold(
-        (failure) => this._readme = null, (data) => this._readme = data);
+        (failure) => this._readme = '', (data) => this._readme = data);
   }
 
   @action
