@@ -1,5 +1,3 @@
-import 'package:flutter/material.dart';
-
 class Score {
   final int grantedPoints;
   final int maxPoints;
@@ -7,27 +5,19 @@ class Score {
   final double popularityScore;
 
   const Score(
-      {@required this.maxPoints,
-      @required this.grantedPoints,
-      @required this.likeCount,
-      @required this.popularityScore});
+      {required this.maxPoints,
+      required this.grantedPoints,
+      required this.likeCount,
+      required this.popularityScore});
 
-  factory Score.fromMap(Map<String, dynamic> map) {
-    return new Score(
-      maxPoints: map['maxPoints'] as int,
-      grantedPoints: map['grantedPoints'] as int,
-      likeCount: map['likeCount'] as int,
-      popularityScore: map['popularityScore'] as double,
-    );
-  }
 
   Map<String, dynamic> toMap() {
     // ignore: unnecessary_cast
     return {
-      'maxPoints': this.maxPoints??0,
-      'likeCount': this.likeCount??0,
-      'popularityScore': this.popularityScore??0,
-      'grantedPoints': this.grantedPoints??0,
+      'maxPoints': this.maxPoints ,
+      'likeCount': this.likeCount ,
+      'popularityScore': this.popularityScore ,
+      'grantedPoints': this.grantedPoints,
     } as Map<String, dynamic>;
   }
 
@@ -37,8 +27,7 @@ class Score {
 
 extension Scorex on Score {
   String getPopularity() {
-    var result =
-        (popularityScore != null) ? (popularityScore * 100).roundToDouble() : 0;
-    return '${result}%';
+    var result = (popularityScore * 100).roundToDouble();
+    return '$result%';
   }
 }
