@@ -1,7 +1,5 @@
 import 'package:dartz/dartz.dart';
 import 'package:dio/dio.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_package/src/domain/packages/entities/package.dart';
 import 'package:flutter_package/src/domain/core/api_service.dart';
 import 'package:flutter_package/src/domain/core/request_failure.dart';
 
@@ -11,5 +9,5 @@ abstract class ISearchService extends ApiService {
   final String path = 'search';
 
   Future<Either<RequestFailure, List<String>>> searchPackage(
-      {@required int page, @required String query});
+      {required int page, required String query});
 }
