@@ -1,21 +1,19 @@
-import 'dart:convert';
-import 'dart:io';
 
+import 'dart:io';
 import 'package:dartz/dartz.dart';
 import 'package:dio/dio.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_package/src/domain/packages/entities/package.dart';
-import 'package:flutter_package/src/domain/core/api_service.dart';
 import 'package:flutter_package/src/domain/core/request_failure.dart';
 import 'package:flutter_package/src/domain/packages/entities/score.dart';
 import 'package:flutter_package/src/domain/packages/i_package_service.dart';
+import 'package:flutter_package/src/infrastructure/core/mapper.dart';
 
 class PackageService extends IPackageService {
   PackageService(Dio dio) : super(dio);
 
   @override
   Future<Either<RequestFailure, List<Package>>> getPackages(
-      {@required int page}) async {
+      {required int page}) async {
     try {
       final response = await dio.get('${path}?page=${page}',
           options: Options(
@@ -28,7 +26,8 @@ class PackageService extends IPackageService {
       if (response.statusCode != 200) return Left(RequestFailure.serverError());
 
       var packages = List.from(response.data['packages']).map<Package>((data) {
-        return Package.fromMap(data);
+        // print('data: ${json.encode(data)}');
+        return Mapper.packageFromMap(data);
       }).toList();
 
       return (packages.isEmpty)
@@ -39,27 +38,27 @@ class PackageService extends IPackageService {
         return left(RequestFailure.networkError());
       }
 
-      if (e.type == DioErrorType.CONNECT_TIMEOUT ||
-          e.type == DioErrorType.RECEIVE_TIMEOUT ||
-          e.type == DioErrorType.SEND_TIMEOUT) {
+      if (e.type == DioErrorType.connectTimeout ||
+          e.type == DioErrorType.receiveTimeout ||
+          e.type == DioErrorType.sendTimeout) {
         return left(RequestFailure.serverError());
       }
 
-      if (e.response.statusCode == 500) {
+      if (e.response == null || e.response!.statusCode == 500) {
         return Left(RequestFailure.serverError());
       }
 
       return Left(RequestFailure.serverError());
     } catch (error) {
+      print('getPackages: $error');
       return left(RequestFailure.serverError());
     }
   }
 
   @override
   Future<Either<RequestFailure, Package>> getPackageName(
-      {String namePackage}) async {
+      {required String namePackage}) async {
     try {
-      print('namePackage: ${namePackage}');
       final response = await dio.get('$path/$namePackage',
           options: Options(
             headers: {
@@ -71,7 +70,7 @@ class PackageService extends IPackageService {
 
       if (response.statusCode != 200) return Left(RequestFailure.serverError());
 
-      var package = Package.from(response.data);
+      var package = Mapper.packageFromMap(response.data);
 
       return Right(package);
     } on DioError catch (e) {
@@ -79,13 +78,13 @@ class PackageService extends IPackageService {
         return left(RequestFailure.networkError());
       }
 
-      if (e.type == DioErrorType.CONNECT_TIMEOUT ||
-          e.type == DioErrorType.RECEIVE_TIMEOUT ||
-          e.type == DioErrorType.SEND_TIMEOUT) {
+      if (e.type == DioErrorType.connectTimeout ||
+          e.type == DioErrorType.receiveTimeout ||
+          e.type == DioErrorType.sendTimeout) {
         return left(RequestFailure.serverError());
       }
 
-      if (e.response.statusCode == 500) {
+      if (e.response == null || e.response!.statusCode == 500) {
         return Left(RequestFailure.serverError());
       }
 
@@ -98,7 +97,7 @@ class PackageService extends IPackageService {
 
   @override
   Future<Either<RequestFailure, Score>> getScorePackage(
-      {String namePackage}) async {
+      {required String namePackage}) async {
     try {
       final response = await dio.get('$path/$namePackage/score',
           options: Options(
@@ -110,7 +109,7 @@ class PackageService extends IPackageService {
 
       if (response.statusCode != 200) return Left(RequestFailure.serverError());
 
-      var score = Score.fromMap(response.data);
+      var score = Mapper.scoreFromMap(response.data);
 
       return Right(score);
     } on DioError catch (e) {
@@ -118,13 +117,13 @@ class PackageService extends IPackageService {
         return left(RequestFailure.networkError());
       }
 
-      if (e.type == DioErrorType.CONNECT_TIMEOUT ||
-          e.type == DioErrorType.RECEIVE_TIMEOUT ||
-          e.type == DioErrorType.SEND_TIMEOUT) {
+      if (e.type == DioErrorType.connectTimeout ||
+          e.type == DioErrorType.receiveTimeout ||
+          e.type == DioErrorType.sendTimeout) {
         return left(RequestFailure.serverError());
       }
 
-      if (e.response.statusCode == 500) {
+      if (e.response == null || e.response!.statusCode == 500) {
         return Left(RequestFailure.serverError());
       }
 
