@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
 abstract class BaseComponent extends StatelessWidget {
-  final String value;
-  BaseComponent({@required this.value});
+  final String? value;
+  BaseComponent({this.value});
 }
