@@ -9,7 +9,7 @@ part of 'base_view_model.dart';
 // ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic
 
 mixin _$BaseViewModel on _BaseViewModel, Store {
-  Computed<bool> _$isBusyComputed;
+  Computed<bool>? _$isBusyComputed;
 
   @override
   bool get isBusy => (_$isBusyComputed ??=
