@@ -1,28 +1,25 @@
-import 'dart:convert';
-
 import 'dependencie.dart';
 import 'environment.dart';
-import 'version.dart';
 
 class Pubspec {
-  String name;
-  String version;
-  String description;
-  String homepage;
-  String repository;
-  Environment environment;
-  List<Dependencie> dependencies;
-  List<Dependencie> dev_dependencies;
+  final String name;
+  final String version;
+  final String description;
+  final String homepage;
+  final String repository;
+  final Environment environment;
+  final List<Dependencie> dependencies;
+  final List<Dependencie> dev_dependencies;
 
   Pubspec(
-      {this.name,
-      this.version,
-      this.description,
-      this.homepage,
-      this.repository,
-      this.environment,
-      this.dependencies,
-      this.dev_dependencies});
+      {required this.name,
+      required this.version,
+      required this.description,
+      required this.homepage,
+      required this.repository,
+      required this.environment,
+      required this.dependencies,
+      required this.dev_dependencies});
 
   Map<String, dynamic> toMap() {
     return {
@@ -36,14 +33,5 @@ class Pubspec {
     };
   }
 
-  factory Pubspec.fromMap(Map<String, dynamic> map) {
 
-    return new Pubspec(
-      name: map['name'] as String,
-      version: map['version'] as String,
-      description: map['description'] as String,
-      repository: map['repository'] as String,
-      homepage: map['homepage'] as String,
-    );
-  }
 }
