@@ -1,11 +1,11 @@
 import 'package:flutter_package/src/domain/packages/entities/pubspec.dart';
 
 class Version {
-  String version;
-  String archive_url;
-  Pubspec pubspec;
+  final String version;
+  final String archive_url;
+  final Pubspec pubspec;
 
-  Version({this.version, this.archive_url, this.pubspec});
+  Version({required this.version, required this.archive_url, required this.pubspec});
 
   Map<String, dynamic> toMap() {
     return {
@@ -15,11 +15,5 @@ class Version {
     };
   }
 
-  factory Version.fromMap(Map<String, dynamic> map) {
-    return new Version(
-      version: map['version'] as String,
-      archive_url: map['archive_url'] as String,
-      pubspec: Pubspec.fromMap(map['pubspec']),
-    );
-  }
+
 }
