@@ -1,12 +1,11 @@
 import 'package:dartz/dartz.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_package/src/domain/core/i_advanced_service.dart';
 import 'package:flutter_package/src/domain/core/request_failure.dart';
 import 'package:flutter_package/src/infrastructure/core/network_builder.dart';
 
 class DefaultAdvancedService extends IAdvancedService {
   @override
-  Future<Either<RequestFailure, String>> getReadme({String gitPath}) async {
+  Future<Either<RequestFailure, String>> getReadme({String? gitPath}) async {
     if (gitPath == null || gitPath.isEmpty)
       return Left(RequestFailure.serverError());
 
