@@ -5,7 +5,7 @@ const String host = 'pub.dartlang.org';
 const String apiBaseUrl = 'https://$host/api/';
 
 class NetworkBuilder {
-  String _baseUrl;
+  String _baseUrl = '';
 
   void setBaseUrl(String baseUrl) {
     ArgumentError.checkNotNull(baseUrl, 'baseUrl');
@@ -14,7 +14,7 @@ class NetworkBuilder {
 
   Dio build() {
     final dio = Dio();
-    dio.options.baseUrl = _baseUrl??apiBaseUrl;
+    dio.options.baseUrl = (_baseUrl.isEmpty) ? apiBaseUrl : _baseUrl;
     final defaultClient = DefaultHttpClientAdapter();
     defaultClient.onHttpClientCreate = (httpClient) {
       httpClient.badCertificateCallback = (_, host, __) => host == apiBaseUrl;
