@@ -1,6 +1,4 @@
 import 'package:dartz/dartz.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter_package/src/domain/packages/entities/package.dart';
 import 'package:flutter_package/src/domain/core/request_failure.dart';
 import 'package:flutter_package/src/domain/search/i_search_repository.dart';
 import 'package:flutter_package/src/domain/search/i_search_service.dart';
@@ -10,6 +8,6 @@ class SearchRepository extends ISearchRepository {
 
   @override
   Future<Either<RequestFailure, List<String>>> searchPackage(
-          {int page, String query}) async =>
+          {required int page, required String query}) async =>
       await service.searchPackage(page: page, query: query);
 }
