@@ -6,7 +6,6 @@ import 'package:flutter_package/src/utils/uihelper.dart';
 import 'package:flutter_package/src/application/packages/packages_view_model.dart';
 import 'package:flutter_package/src/domain/packages/i_package_repository.dart';
 import 'package:flutter_package/src/injection/injection_config.dart';
-import 'package:flutter_package/src/presentation/core/custom_progress.dart';
 import 'package:flutter_package/src/presentation/core/custom_refresh.dart';
 import 'package:flutter_package/src/presentation/core/svg_icon.dart';
 import 'package:flutter_package/src/utils/theme.dart';
@@ -43,7 +42,7 @@ class PackagesScreenState extends State<PackagesScreen>
           crossAxisAlignment: CrossAxisAlignment.center,
           children: <Widget>[
             Txt(
-              _message.translate,
+              _message().translate,
               textAlign: TextAlign.center,
               textColor: Colors.black,
             ),
@@ -65,10 +64,14 @@ class PackagesScreenState extends State<PackagesScreen>
     );
   }
 
-  String get _message => _model.failure.when<String>(
-      networkError: () => 'no_internet_access',
-      empty: () => 'no_results_found',
-      serverError: () => 'server_failure');
+  String   _message (){
+    if(!_model.hasError) return '';
+
+    return _model.failure!.when<String>(
+        networkError: () => 'no_internet_access',
+        empty: () => 'no_results_found',
+        serverError: () => 'server_failure');
+  }
 
   _build() {
     return Observer(builder: (_) {
